feat: Add Google Maps Flutter dependency and implement dispatcher home screen tests
- Added `google_maps_flutter` dependency to `pubspec.yaml`.
- Enhanced `dispatcher_home_screen_test.dart` with shimmer loading state and error handling.
- Created `forgot_password_flow_test.dart` to cover forgot and reset password flows.
- Implemented tests for dispatcher home data mappers and repository.
- Added view model tests for dispatcher home functionality.
- Developed widget tests for dispatcher home driver marker.

+++ lib/features/dispatcher/dispatcher_home/domain/entities/dispatcher_home_operations_status_entity.dart
@@ -11,7 +11,7 @@ class DispatcherHomeOperationsStatusEntity {
     required this.cancelledLabel,
   });
 
-  final int completionRate;
+  final double completionRate;
   final int deliveredCount;
   final String deliveredLabel;
   final int inDeliveryCount;
